Zip file contains the whole code and story board for Unit Converter iOS application. Also attached a video where you can see the execution in iphone 11 pro max simulator
It is a multi-view iOS application where the user selects a choice from tab’s available and then selects the source and destination units from cells available in a table view, which navigates the user to the Input screen where he also gets a validated output and errors if occur.
Bascially this app will calculate and converts data from one unit system to another unit system.